- fixed some potential dangerous stuff

+++ src/common/CXPenHolder.cpp
@@ -21,7 +21,6 @@
  ***************************************************************************/
 
 #include "CXPenHolder.hpp"
-#include "CXPen.hpp"
 
 const int MAXPENWIDTH = 19;
 
@@ -38,7 +37,7 @@ CXPenHolder::~CXPenHolder() {
 //-------------------------------------
 void CXPenHolder::CreatePens() {
 
-    for(size_t i=0; i<e_Way_EnumCount; i++) {
+    for(size_t i=0; i<static_cast<size_t>(e_Way_EnumCount); i++) {
         m_Pens.Append(new CXPens());
     }
 
@@ -46,129 +45,129 @@ void CXPenHolder::CreatePens() {
 
     // now create pens
     // Mapnik
-    pPens = m_Pens[e_Way_Fading];
+    pPens = m_Pens[static_cast<size_t>(e_Way_Fading)];
     pPens->m_pBg        = new CXPen(CXPen::e_Solid, 1, CXRGB(0xB0, 0xB0, 0xB0));
     pPens->m_pSegm      = NULL;
 
-    pPens = m_Pens[e_Way_Unknown];
+    pPens = m_Pens[static_cast<size_t>(e_Way_Unknown)];
     pPens->m_pBg        = new CXPen(CXPen::e_Solid, 1, CXRGB(0x00, 0x00, 0x00));
     pPens->m_pSegm      = NULL;
 
-    pPens = m_Pens[e_Way_Motorway];
+    pPens = m_Pens[static_cast<size_t>(e_Way_Motorway)];
     pPens->m_pBg        = new CXPen(CXPen::e_Solid, 7, CXRGB(0x50, 0x60, 0x76));
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 5, CXRGB(0x80, 0x9B, 0xC0));
 
-    pPens = m_Pens[e_Way_MotorwayLink];
+    pPens = m_Pens[static_cast<size_t>(e_Way_MotorwayLink)];
     pPens->m_pBg        = new CXPen(CXPen::e_Solid, 7, CXRGB(0x50, 0x60, 0x76));
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 5, CXRGB(0x80, 0x9B, 0xC0));
 
-    pPens = m_Pens[e_Way_Trunk];
+    pPens = m_Pens[static_cast<size_t>(e_Way_Trunk)];
     pPens->m_pBg        = new CXPen(CXPen::e_Solid, 7, CXRGB(0x47, 0x71, 0x47));
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 5, CXRGB(0x7E, 0xC8, 0x7E));
 
-    pPens = m_Pens[e_Way_TrunkLink];
+    pPens = m_Pens[static_cast<size_t>(e_Way_TrunkLink)];
     pPens->m_pBg        = new CXPen(CXPen::e_Solid, 7, CXRGB(0x47, 0x71, 0x47));
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 5, CXRGB(0x7E, 0xC8, 0x7E));
 
-    pPens = m_Pens[e_Way_Primary];
+    pPens = m_Pens[static_cast<size_t>(e_Way_Primary)];
     pPens->m_pBg        = new CXPen(CXPen::e_Solid, 5, CXRGB(0x92, 0x46, 0x49));
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 3, CXRGB(0xE3, 0x60, 0x70));
 
-    pPens = m_Pens[e_Way_PrimaryLink];
+    pPens = m_Pens[static_cast<size_t>(e_Way_PrimaryLink)];
     pPens->m_pBg        = new CXPen(CXPen::e_Solid, 5, CXRGB(0x92, 0x46, 0x49));
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 3, CXRGB(0xE3, 0x6C, 0x70));
 
-    pPens = m_Pens[e_Way_Secondary];
+    pPens = m_Pens[static_cast<size_t>(e_Way_Secondary)];
     pPens->m_pBg        = new CXPen(CXPen::e_Solid, 5, CXRGB(0xA5, 0x7D, 0x49));
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 3, CXRGB(0xFC, 0xBE, 0x6E));
 
-    pPens = m_Pens[e_Way_Tertiary];
+    pPens = m_Pens[static_cast<size_t>(e_Way_Tertiary)];
     pPens->m_pBg        = new CXPen(CXPen::e_Solid, 5, CXRGB(0xAA, 0xAA, 0xAA));
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 3, CXRGB(0xFE, 0xFE, 0xCB));
 
-    pPens = m_Pens[e_Way_Unclassified];
+    pPens = m_Pens[static_cast<size_t>(e_Way_Unclassified)];
     pPens->m_pBg        = new CXPen(CXPen::e_Solid, 5, CXRGB(0xAA, 0xAA, 0xAA));
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 3, CXRGB(0xFE, 0xFE, 0xDD));
 
-    pPens = m_Pens[e_Way_Track];
+    pPens = m_Pens[static_cast<size_t>(e_Way_Track)];
     pPens->m_pBg        = NULL;
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 2, CXRGB(0xD7, 0x93, 0x31));
 
-    pPens = m_Pens[e_Way_Residential];
+    pPens = m_Pens[static_cast<size_t>(e_Way_Residential)];
     pPens->m_pBg        = new CXPen(CXPen::e_Solid, 5, CXRGB(0xAA, 0xAA, 0xAA));
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 3, CXRGB(0xFE, 0xFE, 0xFE));
 
-    pPens = m_Pens[e_Way_Service];
+    pPens = m_Pens[static_cast<size_t>(e_Way_Service)];
     pPens->m_pBg        = new CXPen(CXPen::e_Solid, 3, CXRGB(0xAA, 0xAA, 0xAA));
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 1, CXRGB(0xE0, 0xE0, 0xE0));
 
-    pPens = m_Pens[e_Way_Bridleway];
+    pPens = m_Pens[static_cast<size_t>(e_Way_Bridleway)];
     /// \todo implement
     pPens->m_pBg        = new CXPen(CXPen::e_Solid, 1, CXRGB(0x00, 0x00, 0x00));
     pPens->m_pSegm      = NULL;
 
-    pPens = m_Pens[e_Way_Cycleway];
+    pPens = m_Pens[static_cast<size_t>(e_Way_Cycleway)];
     pPens->m_pBg        = NULL;
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 1, CXRGB(0x00, 0x80, 0x00));
 
-    pPens = m_Pens[e_Way_Footway];
+    pPens = m_Pens[static_cast<size_t>(e_Way_Footway)];
     pPens->m_pBg        = NULL;
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 3, CXRGB(0xD7, 0x93, 0x31));
 
-    pPens = m_Pens[e_Way_Pedestrian];
+    pPens = m_Pens[static_cast<size_t>(e_Way_Pedestrian)];
     /// \todo implement
     pPens->m_pBg        = new CXPen(CXPen::e_Solid, 2, CXRGB(0x00, 0x00, 0x00));
     pPens->m_pSegm      = NULL;
 
-    pPens = m_Pens[e_Way_Steps];
+    pPens = m_Pens[static_cast<size_t>(e_Way_Steps)];
     /// \todo implement
     pPens->m_pBg        = new CXPen(CXPen::e_Solid, 1, CXRGB(0x00, 0x00, 0x00));
     pPens->m_pSegm      = NULL;
 
-    pPens = m_Pens[e_Way_LivingStreet];
+    pPens = m_Pens[static_cast<size_t>(e_Way_LivingStreet)];
     pPens->m_pBg        = new CXPen(CXPen::e_Solid, 5, CXRGB(0xAA, 0xAA, 0xAA));
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 3, CXRGB(0xD0, 0xD0, 0xD0));
 
-    pPens = m_Pens[e_Way_Railway_Thick];
+    pPens = m_Pens[static_cast<size_t>(e_Way_Railway_Thick)];
     pPens->m_pBg        = NULL;
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 4, CXRGB(0x50, 0x50, 0x50));
 
-    pPens = m_Pens[e_Way_Railway_Thin];
+    pPens = m_Pens[static_cast<size_t>(e_Way_Railway_Thin)];
     pPens->m_pBg        = NULL;
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 2, CXRGB(0x50, 0x50, 0x50));
 
-    pPens = m_Pens[e_Way_Water_Thick];
+    pPens = m_Pens[static_cast<size_t>(e_Way_Water_Thick)];
     pPens->m_pBg        = NULL;
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 4, CXRGB(0x00, 0x00, 0xD0));
 
-    pPens = m_Pens[e_Way_Water_Thin];
+    pPens = m_Pens[static_cast<size_t>(e_Way_Water_Thin)];
     pPens->m_pBg        = NULL;
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 2, CXRGB(0x00, 0x00, 0xD0));
 
-    pPens = m_Pens[e_Way_BorderThick];
+    pPens = m_Pens[static_cast<size_t>(e_Way_BorderThick)];
     pPens->m_pBg        = NULL;
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 5, CXRGB(0xDD, 0x00, 0x00));
 
-    pPens = m_Pens[e_Way_BorderMedium];
+    pPens = m_Pens[static_cast<size_t>(e_Way_BorderMedium)];
     pPens->m_pBg        = NULL;
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 3, CXRGB(0xDD, 0x00, 0x00));
 
-    pPens = m_Pens[e_Way_BorderThin];
+    pPens = m_Pens[static_cast<size_t>(e_Way_BorderThin)];
     pPens->m_pBg        = NULL;
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 1, CXRGB(0xDD, 0x00, 0x00));
 
-    pPens = m_Pens[e_Way_Path];
+    pPens = m_Pens[static_cast<size_t>(e_Way_Path)];
     pPens->m_pBg        = NULL;
     pPens->m_pSegm      = new CXPen(CXPen::e_Solid, 1, CXRGB(0xD7, 0x93, 0x31));
 
-    for(size_t j=0; j<e_Way_EnumCount; j++) {
+    for(size_t j=0; j<static_cast<size_t>(e_Way_EnumCount); j++) {
         m_ScaledPens.Append(new CXPens(*m_Pens[j]));
     }
 }
 
 //-------------------------------------
 void CXPenHolder::DestroyPens() {
-    for(size_t i=0; i<e_Way_EnumCount; i++) {
+    for(size_t i=0; i<static_cast<size_t>(e_Way_EnumCount); i++) {
         CXPens *pPens = m_Pens[i];
         delete pPens;
         pPens = m_ScaledPens[i];
@@ -180,20 +179,20 @@ void CXPenHolder::DestroyPens() {
 
 //-------------------------------------
 CXPen *CXPenHolder::GetPenBg(E_WAY_TYPE eWayType) {
-    CXPen *Result = m_Pens[eWayType]->m_pBg;
+    CXPen *Result = m_Pens[static_cast<size_t>(eWayType)]->m_pBg;
     return Result;
 }
 
 //-------------------------------------
 CXPen *CXPenHolder::GetPenFg(E_WAY_TYPE eWayType) {
-    CXPen *Result = m_Pens[eWayType]->m_pSegm;
+    CXPen *Result = m_Pens[static_cast<size_t>(eWayType)]->m_pSegm;
     return Result;
 }
 
 //-------------------------------------
 void CXPenHolder::ScalePens(double ScaleFactor) {
-    for(size_t i=0; i<e_Way_EnumCount; i++) {
-        if(i == e_Way_Unknown)
+    for(size_t i=0; i<static_cast<size_t>(e_Way_EnumCount); i++) {
+        if(i == static_cast<size_t>(e_Way_Unknown))
             // do not scale pen for way of type e_Way_Unknown
             continue;
         CXPens *pPens = m_Pens[i];
@@ -225,12 +224,12 @@ void CXPenHolder::ScalePens(double ScaleFactor) {
 
 //-------------------------------------
 CXPen *CXPenHolder::GetScaledPenBg(E_WAY_TYPE eWayType) {
-    CXPen *Result = m_ScaledPens[eWayType]->m_pBg;
+    CXPen *Result = m_ScaledPens[static_cast<size_t>(eWayType)]->m_pBg;
     return Result;
 }
 
 //-------------------------------------
 CXPen *CXPenHolder::GetScaledPenFg(E_WAY_TYPE eWayType) {
-    CXPen *Result = m_ScaledPens[eWayType]->m_pSegm;
+    CXPen *Result = m_ScaledPens[static_cast<size_t>(eWayType)]->m_pSegm;
     return Result;
 }
